Добавил маршрутизацию, вынес страницы, продолжил работу с компонентами
