+++ project/src/const/enums.ts
@@ -1,0 +1,13 @@
+export enum AppRoute {
+  Root = '/',
+  Offer = '/offer/:id',
+  SignIn = '/login',
+  Favorites = 'favorites',
+  NotFound = '*',
+}
+
+export enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN',
+}
