+++ project/src/pages/favorites/favorites.tsx
@@ -1,49 +1,9 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../const/enums';
+
 function Favorites() {
   return (
-    <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a
-                className="header__logo-link"
-                href="main.html"
-              >
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width="81"
-                  height="41"
-                />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a
-                    className="header__nav-link"
-                    href="#"
-                  >
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+    <>
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
@@ -218,10 +178,11 @@ function Favorites() {
           </section>
         </div>
       </main>
+
       <footer className="footer container">
-        <a
+        <Link
           className="footer__logo-link"
-          href="main.html"
+          to={AppRoute.Root}
         >
           <img
             className="footer__logo"
@@ -230,9 +191,9 @@ function Favorites() {
             width="64"
             height="33"
           />
-        </a>
+        </Link>
       </footer>
-    </div>
+    </>
   );
 }
