+++ project/src/pages/sign-in/sign-in.tsx
@@ -1,0 +1,28 @@
+import SignInForm from '../../components/sign-in-form/sign-in-form';
+
+function SignIn() {
+  return (
+    <main className="page__main page__main--login">
+      <div className="page__login-container container">
+        <section className="login">
+          <h1 className="login__title">Sign in</h1>
+
+          <SignInForm />
+        </section>
+
+        <section className="locations locations--login locations--current">
+          <div className="locations__item">
+            <a
+              className="locations__item-link"
+              href="#"
+            >
+              <span>Amsterdam</span>
+            </a>
+          </div>
+        </section>
+      </div>
+    </main>
+  );
+}
+
+export default SignIn;
